requirements.txt
numpy>=1.24
pandas>=2.0
requests>=2.31
schedule>=1.2
python-dotenv>=1.0
python-telegram-bot>=21.0
xgboost>=2.0
ccxt>=4.0
